requirements.txt
# 🔥 ARBI PHOENIX - Requirements

# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Trading APIs
MetaTrader5>=5.0.45
python-ib-api>=9.81.1

# Data Processing & Analysis
scikit-learn>=1.0.0
statsmodels>=0.13.0
ta-lib>=0.4.24

# GUI Framework
PyQt6>=6.4.0
qasync>=0.24.0
matplotlib>=3.5.0
plotly>=5.10.0

# Async & Networking
asyncio-mqtt>=0.11.0
websocket-client>=1.4.0
aiohttp>=3.8.0
requests>=2.28.0

# Database
sqlite3
sqlalchemy>=1.4.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0
loguru>=0.6.0
schedule>=1.1.0

# Notifications
plyer>=2.1.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
flake8>=5.0.0

# Performance
numba>=0.56.0
cython>=0.29.0

# Optional: For advanced features
tensorflow>=2.10.0  # For ML predictions
torch>=1.12.0       # Alternative ML framework
